Return response entity and correct status codes

+++ src/main/java/org/aitesting/passenger/controllers/PassengerController.java
@@ -1,66 +1,59 @@
 package org.aitesting.passenger.controllers;
 
 import org.aitesting.passenger.model.Passenger;
-import org.aitesting.passenger.model.PassengerDto;
 import org.aitesting.passenger.services.PassengerService;
-import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import java.util.List;
+
+import javax.persistence.EntityNotFoundException;
 
 @RestController
-@RequestMapping("api/v1")
+@RequestMapping("api/v1/passenger")
 public class PassengerController {
 
 
     @Autowired
     private PassengerService passengerService;
 
-    private ModelMapper modelMapper = new ModelMapper();
 
-
-    @GetMapping(value="passenger")
-    public List<Passenger> getAllPassengers() {
-        return passengerService.getList();
+    @GetMapping
+    public ResponseEntity getAllPassengers() throws EntityNotFoundException {
+        return new ResponseEntity<>(passengerService.getList(),HttpStatus.OK);
     }
 
-    @GetMapping(value="passenger/{passID}")
-    public Passenger getPassenger(@PathVariable long passID) {
-        return passengerService.getPassenger(passID);
+    @GetMapping(value="{passID}")
+    public ResponseEntity getPassenger(@PathVariable long passID) throws EntityNotFoundException {
+        if(!passengerService.isPassengerById(passID)) {
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(passengerService.getPassenger(passID),HttpStatus.OK);
     }
 
-    @PostMapping(value = "passenger")
-    public ResponseEntity addPassenger(@RequestBody PassengerDto passengerDto){
-        try {
-            Passenger passenger = convertToPassenger(passengerDto);
+    @PostMapping
+    public ResponseEntity addPassenger(@RequestBody Passenger passenger){
             passengerService.addPassenger(passenger);
-            return ResponseEntity.status(HttpStatus.CREATED)
-                    .body(passenger);
-        }
-        catch(Exception e){
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body("Passenger cannot be created");
-        }
+            return new ResponseEntity<>(passenger, HttpStatus.CREATED);
     }
 
-    @PutMapping(value="passenger")
-    public void updatePassenger(@RequestBody Passenger passenger){
-        passengerService.updatePassenger(passenger);
+    @PutMapping
+    public ResponseEntity updatePassenger(@RequestBody Passenger passenger) throws EntityNotFoundException {
+            if(!passengerService.isPassenger(passenger)) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+
+            passengerService.updatePassenger(passenger);
+            return new ResponseEntity<>("Updated",HttpStatus.OK);
     }
 
-    @DeleteMapping(value = "passenger/{passID}")
-    public void deletePassenger(@PathVariable long passID){
+    @DeleteMapping(value = "{passID}")
+    public ResponseEntity deletePassenger(@PathVariable long passID) throws EntityNotFoundException {
+        if(!passengerService.isPassengerById(passID)){
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
         passengerService.deletePassenger(passID);
-    }
-
-    private Passenger convertToPassenger(PassengerDto passengerDto) {
-        return modelMapper.map(passengerDto, Passenger.class);
-    }
-
-    private PassengerDto convertToPassengerDto(Passenger passenger) {
-        return modelMapper.map(passenger, PassengerDto.class);
+        return new ResponseEntity(HttpStatus.OK);
     }
 
 }
